[clang][ssaf][NFC] Hoist findFnByName and findDeclByName
Split from #194448

This was already approved in
https://github.com/llvm/llvm-project/pull/194448#pullrequestreview-4201251523
